style: update style
