+++ src/components/SearchBar/components/AdvancedInput.tsx
@@ -21,10 +21,10 @@ type AdvancedFormValues = {
   "Sitter has a fully fenced backyard": boolean;
 };
 
-type AdvancedFormCheckBoxNames =
-  | "Sitter has no dogs"
-  | "Sitter has no children"
-  | "Sitter has a fully fenced backyard";
+// type AdvancedFormCheckBoxNames =
+//   | "Sitter has no dogs"
+//   | "Sitter has no children"
+//   | "Sitter has a fully fenced backyard";
 
 const AdvancedCheckItemText = chakra(Text, {
   baseStyle: {
@@ -35,16 +35,24 @@ const AdvancedCheckItemText = chakra(Text, {
   },
 });
 
-const advancedCheckList = [
-  { id: 1, content: "Sitter has no dogs" },
-  { id: 2, content: "Sitter has no children" },
-  { id: 3, content: "Sitter has a fully fenced backyard" },
-];
+// const advancedCheckList = [
+//   { id: 1, content: "Sitter has no dogs" },
+//   { id: 2, content: "Sitter has no children" },
+//   { id: 3, content: "Sitter has a fully fenced backyard" },
+// ];
 
 const AdvancedInput = () => {
-  const [advancedText, setAdvancedText] = useState(<Box>Advanced filters</Box>);
-  const [isActive, setIsActive] = useState(false);
+  const [isChecked1, setIsChecked1] = useState(false);
+  const [isChecked2, setIsChecked2] = useState(false);
+  const [isChecked3, setIsChecked3] = useState(false);
   const [activeItemsCount, setActiveItemsCount] = useState(0);
+  const [activeItemsCount1, setActiveItemsCount1] = useState(0);
+  const [activeItemsCount2, setActiveItemsCount2] = useState(0);
+  const [activeItemsCount3, setActiveItemsCount3] = useState(0);
+
+  useEffect(() => {
+    setActiveItemsCount(activeItemsCount1 + activeItemsCount2 + activeItemsCount3);
+  });
 
   // const handleText = () => {
   //   setIsActive(!isActive);
@@ -65,6 +73,63 @@ const AdvancedInput = () => {
   //   }
   // }, [isActive, setActiveItemsCount]);
 
+  const handleText1 = () => {
+    setIsChecked1(!isChecked1);
+  };
+  useEffect(() => {
+    if (!isChecked1) {
+      setActiveItemsCount1((prevCount) => {
+        if (prevCount !== 0) {
+          return prevCount - 1;
+        }
+        return prevCount;
+      });
+    }
+    if (isChecked1) {
+      setActiveItemsCount1((prevCount) => {
+        return prevCount + 1;
+      });
+    }
+  }, [isChecked1, setActiveItemsCount1]);
+
+  const handleText2 = () => {
+    setIsChecked2(!isChecked2);
+  };
+  useEffect(() => {
+    if (!isChecked2) {
+      setActiveItemsCount2((prevCount) => {
+        if (prevCount !== 0) {
+          return prevCount - 1;
+        }
+        return prevCount;
+      });
+    }
+    if (isChecked2) {
+      setActiveItemsCount2((prevCount) => {
+        return prevCount + 1;
+      });
+    }
+  }, [isChecked2, setActiveItemsCount2]);
+
+  const handleText3 = () => {
+    setIsChecked3(!isChecked3);
+  };
+  useEffect(() => {
+    if (!isChecked3) {
+      setActiveItemsCount3((prevCount) => {
+        if (prevCount !== 0) {
+          return prevCount - 1;
+        }
+        return prevCount;
+      });
+    }
+    if (isChecked3) {
+      setActiveItemsCount3((prevCount) => {
+        return prevCount + 1;
+      });
+    }
+  }, [isChecked3, setActiveItemsCount3]);
+
   const {
     handleSubmit: handleAdvanceInputSubmit,
     control,
@@ -79,6 +144,12 @@ const AdvancedInput = () => {
 
   const handleAdvancedFormReset = () => {
     reset();
+    setActiveItemsCount1(0);
+    setActiveItemsCount2(0);
+    setActiveItemsCount3(0);
+    setIsChecked1(false);
+    setIsChecked2(false);
+    setIsChecked3(false);
   };
 
   const onAdvancedSubmit: SubmitHandler<AdvancedFormValues> = (data) => {
@@ -99,8 +170,7 @@ const AdvancedInput = () => {
           >
             <Box display="flex" alignItems="center" justifyContent="space-around">
               <Box color="rgb(116, 116, 116)">
-                {advancedText}
-                {activeItemsCount}
+                <Box>{activeItemsCount} advanced filters</Box>
               </Box>
               <Icon as={MdArrowDropDown} fontSize="24px" color="rgb(116, 116, 116)" />
             </Box>
@@ -112,7 +182,7 @@ const AdvancedInput = () => {
               </Text>
               <Box>
                 <Stack spacing="1rem">
-                  {advancedCheckList.map((item) => {
+                  {/* {advancedCheckList.map((item) => {
                     return (
                       <Controller
                         control={control}
@@ -144,7 +214,78 @@ const AdvancedInput = () => {
                         )}
                       />
                     );
-                  })}
+                  })} */}
+                  <Controller
+                    control={control}
+                    name="Sitter has no dogs"
+                    defaultValue={false}
+                    render={({ field: { onChange, value, ref } }) => (
+                      <Checkbox
+                        margin="0"
+                        size="lg"
+                        colorScheme="white"
+                        iconColor="rgb(0, 195, 138)"
+                        sx={{ "> span: first-of-type": { boxShadow: "unset" } }}
+                        _hover={{ borderColor: "rgb(0, 195, 138)" }}
+                        onChange={(e) => {
+                          onChange(e);
+                          handleText1();
+                        }}
+                        ref={ref}
+                        isChecked={value}
+                      >
+                        <AdvancedCheckItemText>Sitter has no dogs</AdvancedCheckItemText>
+                      </Checkbox>
+                    )}
+                  />
+                  <Controller
+                    control={control}
+                    name="Sitter has no children"
+                    defaultValue={false}
+                    render={({ field: { onChange, value, ref } }) => (
+                      <Checkbox
+                        margin="0"
+                        size="lg"
+                        colorScheme="white"
+                        iconColor="rgb(0, 195, 138)"
+                        sx={{ "> span: first-of-type": { boxShadow: "unset" } }}
+                        _hover={{ borderColor: "rgb(0, 195, 138)" }}
+                        onChange={(e) => {
+                          onChange(e);
+                          handleText2();
+                        }}
+                        ref={ref}
+                        isChecked={value}
+                      >
+                        <AdvancedCheckItemText>Sitter has no children</AdvancedCheckItemText>
+                      </Checkbox>
+                    )}
+                  />
+                  <Controller
+                    control={control}
+                    name="Sitter has a fully fenced backyard"
+                    defaultValue={false}
+                    render={({ field: { onChange, value, ref } }) => (
+                      <Checkbox
+                        margin="0"
+                        size="lg"
+                        colorScheme="white"
+                        iconColor="rgb(0, 195, 138)"
+                        sx={{ "> span: first-of-type": { boxShadow: "unset" } }}
+                        _hover={{ borderColor: "rgb(0, 195, 138)" }}
+                        onChange={(e) => {
+                          onChange(e);
+                          handleText3();
+                        }}
+                        ref={ref}
+                        isChecked={value}
+                      >
+                        <AdvancedCheckItemText>
+                          Sitter has a fully fenced backyard
+                        </AdvancedCheckItemText>
+                      </Checkbox>
+                    )}
+                  />
                 </Stack>
               </Box>
               <Box display="flex" justifyContent="space-between" paddingTop="1rem">
